Split realtime v1/v2 logic into isolated modules
