+++ codex-rs/codex-api/src/endpoint/realtime_websocket/mod.rs
@@ -1,5 +1,9 @@
 pub mod methods;
-pub mod protocol;
+mod mode_v1;
+mod mode_v2;
+mod protocol_v1;
+mod protocol_v2;
+mod types;
 
 pub use codex_protocol::protocol::RealtimeAudioFrame;
 pub use codex_protocol::protocol::RealtimeEvent;
@@ -7,5 +11,5 @@ pub use methods::RealtimeWebsocketClient;
 pub use methods::RealtimeWebsocketConnection;
 pub use methods::RealtimeWebsocketEvents;
 pub use methods::RealtimeWebsocketWriter;
-pub use protocol::RealtimeApiMode;
-pub use protocol::RealtimeSessionConfig;
+pub use types::RealtimeApiMode;
+pub use types::RealtimeSessionConfig;
